Move FakeProvider and related tests under creds/credstest.
